Add lite mode for less RAM/CPUs
I am on a device with 16GB RAM and unable
to run as configured. Thats difficult to work with
since I'm constantly stashing/popping changes to the
vagrantfile if I'm also trying to stay up to date
with the latest version.

This allows me to just set an env var and dev
as anyone else would.

Signed-off-by: John Schnake <jschnake@vmware.com>