Force checkout of k8s
To avoid re-cloning all of k8s I usually copy over
my repo from my normal path. But when I do that and am
on `master` I get warnings when trying to checkout the
given sha.

Since this is a kubernetes folder just for this project,
it seems reasonable to me to add the `-f` checkout and
discard local changes if they exist.

Signed-off-by: John Schnake <jschnake@vmware.com>