Tweak readme to avoid confusion over vagrant up
The step to bring the cluster up is `make all` or
`make 2-vagrant-up`. The top of the readme makes this
clear but later on it is stated that all you have to do
to bring up the cluster is `vagrant up`.

This may have been the case in the past but now you have
to run `make 2-vagrant-up` to ensure all the proper steps
are followed to not just create the machines but also
initialize the appropriate CNI.

Signed-off-by: John Schnake <jschnake@vmware.com>