Clean up building from source
The build from source option is typically used when developing code
which is then deployed to the VMs. Given that it doesn't make sense to
not build just because the _output directory is present. Also when
building from source there is no reason to compute KUBERNETES_VERSION.

Stop setting path by default in the makefile as that needs to be passed
explicitly by the user.