Add Magefiles as port of Makefile functionality
This port is supposed to offer functional equivalent of the existing
Makefile, but its implementation is also inspired by make.ps1
recently proposed in #254. If the Magefile is preferred and accepted,
then it is supposed to replace make.ps1 without merging of #254.

This is first stab at the idea of the true Windows support in #203