test(cli): Start guide on testing Windows VM on Windows host
This is a draft of a simple step-by-step guide for those
who want to try and test the SWDT CLI on Windows host
targetting Windows VM using PowerShell - arbitrary choice
of native shell on Windows - in order to set it up for a
cluster node (as SWDT CLI is being completed with features).

Although I called the sample `mloskot`, it is written in
a way that everyone should be able to follow it as it is
- it includes reasonable defaults - or follow it to write
own config specific to local environment.
I use this sample during SWDT CLI development and testing,
so it would be helpful if it is hosted within the codebase.

This sample also serves purpose of runnable user stories,
so it may be useful for reviewers to evalute the overall
direction of the SWDT CLI development.

Signed-off-by: Mateusz Łoskot <mateusz@loskot.net>