fix(cli): session.Run is sensitive to quoting on Windows host
There is a strange issue with SSH executor via session.Run
on Windows host (i.e. Windows SSH client) which seems to be
prone to failure depending on particular arrangements of quote,
double-quotes, etc. in PowerShell command.

For example, this fails

  output, err := r.run(`Enable-NetFirewallRule -DisplayGroup "Remote Desktop"`)

while this succeeds

  output, err := r.run(`Enable-NetFirewallRule -DisplayGroup 'Remote Desktop'`)

Wrapping the command with { and } also seems to solve
this problem.

Interestingly, this issue seems not reproducible Linux host.